refactor: remove dead code and tighten visibility in client crate
The `required` module contained feature spec definitions and version
tuples that were never consumed — this information is authoritatively
tracked in the `databend-meta-version` crate instead. Remove it along
with its public re-exports (`FeatureSpec`, `VersionTuple`).

Also remove unnecessary top-level re-exports of internal types
(`GetKVReply`, `ListKVReply`, `MGetKVReply`, `UpsertKVReply`,
`ClientWorkerRequest`, `InitFlag`) and narrow `endpoints` visibility
from `pub` to `pub(crate)`.
