+++ crates/client/src/client_handle.rs
@@ -46,18 +46,18 @@ use tokio::sync::mpsc::UnboundedSender;
 use tokio::sync::oneshot;
 use tonic::codegen::BoxStream;
 
-use crate::ClientWorkerRequest;
-use crate::InitFlag;
-use crate::RequestFor;
-use crate::Streamed;
 use crate::established_client::EstablishedClient;
 use crate::grpc_action::GetKVReply;
 use crate::grpc_action::ListKVReq;
 use crate::grpc_action::MGetKVReply;
 use crate::grpc_action::MGetKVReq;
+use crate::grpc_action::RequestFor;
 use crate::grpc_action::UpsertKVReply;
 use crate::message;
+use crate::message::ClientWorkerRequest;
+use crate::message::InitFlag;
 use crate::message::Response;
+use crate::message::Streamed;
 
 /// A handle to access meta-client worker.
 /// The worker will be actually running in a dedicated runtime: `MetaGrpcClient.rt`.
