+++ crates/client/src/grpc_client.rs
@@ -75,16 +75,16 @@ use tonic::service::Interceptor;
 use tonic::transport::Channel;
 
 use crate::ClientHandle;
-use crate::ClientWorkerRequest;
-use crate::MetaChannelManager;
-use crate::MetaGrpcReadReq;
+use crate::channel_manager::MetaChannelManager;
 use crate::client_conf::RpcClientConf;
 use crate::client_conf::RpcClientTlsConfig;
 use crate::endpoints::Endpoints;
 use crate::endpoints::rotate_failing_endpoint;
 use crate::errors::CreationError;
 use crate::established_client::EstablishedClient;
+use crate::grpc_action::MetaGrpcReadReq;
 use crate::message;
+use crate::message::ClientWorkerRequest;
 use crate::message::Response;
 use crate::pool::Pool;
 use crate::rpc_handler::ResponseAction;
