+++ crates/client/src/lib.rs
@@ -27,14 +27,13 @@ pub extern crate databend_meta_version as version;
 mod channel_manager;
 mod client_conf;
 mod client_handle;
-pub mod endpoints;
+pub(crate) mod endpoints;
 pub mod errors;
 pub(crate) mod established_client;
 mod grpc_action;
 mod grpc_client;
 mod message;
 mod pool;
-pub mod required;
 pub(crate) mod rpc_handler;
 
 pub use channel_manager::DEFAULT_GRPC_MESSAGE_SIZE;
@@ -43,20 +42,12 @@ pub use client_conf::RpcClientConf;
 pub use client_conf::RpcClientTlsConfig;
 pub use client_handle::ClientHandle;
 pub use databend_meta_version::MIN_SERVER_VERSION;
-pub use grpc_action::GetKVReply;
 pub use grpc_action::GetKVReq;
-pub use grpc_action::ListKVReply;
 pub use grpc_action::ListKVReq;
-pub use grpc_action::MGetKVReply;
 pub use grpc_action::MGetKVReq;
 pub use grpc_action::MetaGrpcReadReq;
 pub use grpc_action::MetaGrpcReq;
 pub use grpc_action::RequestFor;
-pub use grpc_action::UpsertKVReply;
 pub use grpc_client::MetaGrpcClient;
 pub use grpc_client::handshake;
-pub use message::ClientWorkerRequest;
-pub use message::InitFlag;
 pub use message::Streamed;
-pub use required::FeatureSpec;
-pub use required::VersionTuple;
